Remove debug verification, use size_t, re-add basic asserts

+++ src/hotspot/share/gc/g1/heapRegion.hpp
@@ -250,7 +250,7 @@ private:
   uint _node_index;
 
   // Number of objects in this region that are currently pinned.
-  volatile uint _pinned_object_count;
+  volatile size_t _pinned_object_count;
 
   void report_region_type_change(G1HeapRegionTraceType::Type to);
